+++ src/renderer/components/inputs/props.ts
@@ -1,11 +1,4 @@
-import {
-    InputData,
-    InputId,
-    InputSchemaValue,
-    OutputId,
-    SchemaId,
-    Size,
-} from '../../../common/common-types';
+import { InputData, InputId, InputSchemaValue, SchemaId, Size } from '../../../common/common-types';
 import { Type } from '../../../common/types/types';
 
 export interface InputProps {
@@ -24,15 +17,4 @@ export interface InputProps {
     readonly useInputSize: (
         inputId: InputId
     ) => readonly [Readonly<Size> | undefined, (size: Readonly<Size>) => void];
-}
-
-export interface OutputProps {
-    readonly id: string;
-    readonly outputId: OutputId;
-    readonly label: string;
-    readonly schemaId: SchemaId;
-    readonly definitionType: Type;
-    readonly hasHandle: boolean;
-    readonly useOutputData: (outputId: OutputId) => unknown;
-    readonly animated: boolean;
 }
